requirements.txt
flask
flask-cors
Pillow
numpy